feat(): refactor types
