+++ packages/plasma-new-hope/src/components/Dropdown/index.ts
@@ -1,5 +1,4 @@
 export { dropdownRoot, dropdownConfig } from './Dropdown';
 export { tokens as dropdownTokens } from './Dropdown.tokens';
 
-export type { DropdownItemOption } from './ui/DropdownItem/DropdownItem.type';
-export type { DropdownProps } from './Dropdown.types';
+export type { DropdownProps, DropdownItemOption } from './Dropdown.types';
