+++ packages/plasma-new-hope/src/components/Dropdown/ui/DropdownInner/DropdownInner.type.ts
@@ -1,7 +1,6 @@
 import type { Dispatch, RefObject } from 'react';
 
-import type { DropdownItemOption } from '../DropdownItem/DropdownItem.type';
-import type { DropdownProps } from '../../Dropdown.types';
+import type { DropdownProps, DropdownItemOption } from '../../Dropdown.types';
 import type { PathState, PathAction } from '../../reducers/pathReducer';
 
 export interface DropdownInnerProps {
